baring bugs which i'm sure there are some...
and the fact i want to clean it up a little more this is a working
version.

middleware on backend checks for accessToken and if it's valid
if so move on to gql
  if not return Not Auth
front end listens for JWTInvalid / expires and trys to hit up refresh
route.
if successful sets new acc to context and localStorage (still need to
  decide on this part) and all continues.

  if at any point there is no accessToken in localStorage or bad res
    comes back we run logout and reload router basically setting you
    back to square one.